Create register a new service with Consul
This code is written for consul open source  and uses the requests library to make a PUT request to the Consul API to register a new service. The service details are defined in a Python dictionary and include the service ID, name, tags, address, and port. The code also defines a health check for the service, which is an HTTP check that pings the service's health endpoint every 10 seconds.

Once the service is registered with Consul, it can be queried and discovered by other services.